redo player damaged state
Remake the player damage state, the old way used the decorator pattern, which makes it not possible for the player class to invoke damage (or decorate) itself. Separated the immunity time frame and the knock back. We can now make player invulnerable without having to take damage, which will be useful for clock item implementation. Knock back is now based on which side of the player is hit, not the player's facing direction.